refactor: extract game loading into minarch_game.c
Move Game_open, Game_close, Game_changeDisc, and extract_zip out of
the monolithic minarch.c into a dedicated minarch_game.c module.
disk_control_ext also moves here since it is owned by game disc management.
