+++ workspace/all/minarch/minarch.c
@@ -44,6 +44,7 @@
 #include "minarch_saves.h"
 #include "minarch_video.h"
 #include "minarch_core.h"
+#include "minarch_game.h"
 
 ///////////////////////////////////////
 
@@ -98,277 +99,6 @@ GFX_Renderer renderer;
 
 struct Core core;
 
-int extract_zip(char** extensions);
-bool getAlias(char* path, char* alias);
-
-struct Game game;
-static void Game_open(char* path) {
-	LOG_info("Game_open\n");
-	int skipzip = 0;
-	memset(&game, 0, sizeof(game));
-	
-	strcpy((char*)game.path, path);
-	strcpy((char*)game.name, strrchr(path, '/')+1);
-	strcpy((char*)game.alt_name, game.name); // default it
-
-	// check first if the rom already is alive in tmp folder if so skip unzipping shit
-	char tmpfldr[255];
-	snprintf(tmpfldr, sizeof(tmpfldr), "/tmp/nextarch/%s", core.tag);
-	char *tmppath = findFileInDir(tmpfldr, game.name);
-	if (tmppath) {
-		// Verify the file exists and has non-zero size (not being written or truncated)
-		struct stat st;
-		if (stat(tmppath, &st) == 0 && st.st_size > 0) {
-			printf("File exists skipping unzipping and setting game.tmp_path: %s\n", tmppath);
-			strcpy((char*)game.tmp_path, tmppath);
-			skipzip = 1;
-			// Update the game name to the extracted file name instead of the zip name
-			if (CFG_getUseExtractedFileName())
-				strcpy((char*)game.alt_name, strrchr(game.tmp_path, '/')+1);
-		} else {
-			printf("File exists but is empty or inaccessible, will re-extract: %s\n", tmppath);
-		}
-		free(tmppath);
-	} else {
-		printf("File does not exist in %s\n",tmpfldr);
-	}
-		
-	// if we have a zip file
-	if (suffixMatch(".zip", game.path) && !skipzip) {
-		LOG_info("is zip file\n");
-		int supports_zip = 0;
-		int i = 0;
-		char* ext;
-		char exts[128];
-		char* extensions[32];
-		strcpy(exts,core.extensions);
-		while ((ext=strtok(i?NULL:exts,"|"))) {
-			extensions[i++] = ext;
-			if (!strcmp("zip", ext)) {
-				supports_zip = 1;
-				break;
-			}
-		}
-		extensions[i] = NULL;
-	
-		// if the core doesn't support zip files natively
-		if (!supports_zip) {
-			// extract zip file located at game.path to game.tmp_path
-			// game.tmp_path is temp dir generated by mkdtemp
-			LOG_info("Extracting zip file manually: %s\n", game.path);
-			if(!extract_zip(extensions))
-				return;
-			// Update the game name to the extracted file name instead of the zip name
-			if (CFG_getUseExtractedFileName())
-				strcpy((char*)game.alt_name, strrchr(game.tmp_path, '/')+1);
-		}
-		else {
-			LOG_info("Core can handle zip file: %s\n", game.path);
-		}
-	}
-		
-	// some cores handle opening files themselves, eg. pcsx_rearmed
-	// if the frontend tries to load a 500MB file itself bad things happen
-	if (!core.need_fullpath) {
-		path = game.tmp_path[0]=='\0'?game.path:game.tmp_path;
-
-		FILE *file = fopen(path, "r");
-		if (file==NULL) {
-			LOG_error("Error opening game: %s\n\t%s\n", path, strerror(errno));
-			return;
-		}
-	
-		fseek(file, 0, SEEK_END);
-		game.size = ftell(file);
-	
-		rewind(file);
-		game.data = malloc(game.size);
-		if (game.data==NULL) {
-			LOG_error("Couldn't allocate memory for file: %s\n", path);
-			return;
-		}
-	
-		fread(game.data, sizeof(uint8_t), game.size, file);
-	
-		fclose(file);
-	}
-	
-	// m3u-based?
-	char* tmp;
-	char m3u_path[256];
-	char base_path[256];
-	char dir_name[256];
-
-	strcpy(m3u_path, game.path);
-	tmp = strrchr(m3u_path, '/') + 1;
-	tmp[0] = '\0';
-	
-	strcpy(base_path, m3u_path);
-	
-	tmp = strrchr(m3u_path, '/');
-	tmp[0] = '\0';
-
-	tmp = strrchr(m3u_path, '/');
-	strcpy(dir_name, tmp);
-	
-	tmp = m3u_path + strlen(m3u_path); 
-	strcpy(tmp, dir_name);
-	
-	tmp = m3u_path + strlen(m3u_path);
-	strcpy(tmp, ".m3u");
-	
-	if (exists(m3u_path)) {
-		strcpy(game.m3u_path, m3u_path);
-		strcpy((char*)game.name, strrchr(m3u_path, '/')+1);
-		strcpy((char*)game.alt_name, game.name); // default it
-	}
-	
-	game.is_open = 1;
-}
-static void Game_close(void) {
-	if (game.data) free(game.data);
-	// why delete tempfile? keep it for next time when loading the game its much faster from /tmp ram folder
-	// if (game.tmp_path[0]) remove(game.tmp_path);
-	game.is_open = 0;
-	VIB_setStrength(0); // just in case
-}
-
-static struct retro_disk_control_ext_callback disk_control_ext;
-static void Game_changeDisc(char* path) {
-	
-	if (exactMatch(game.path, path) || !exists(path)) return;
-	
-	Game_close();
-	Game_open(path);
-	
-	struct retro_game_info game_info = {};
-	game_info.path = game.path;
-	game_info.data = game.data;
-	game_info.size = game.size;
-	
-	disk_control_ext.replace_image_index(0, &game_info);
-	putFile(CHANGE_DISC_PATH, path); // NextUI still needs to know this to update recents.txt
-}
-
-int extract_zip(char** extensions)
-{
-	char buf[100];
-	struct zip *za;
-	int ze;
-	if ((za = zip_open(game.path, 0, &ze)) == NULL) {
-		zip_error_t error;
-		zip_error_init_with_code(&error, ze);
-		LOG_error("can't open zip archive `%s': %s\n", game.path, zip_error_strerror(&error));
-		return 0;
-	}
-
-	// char tmp_template[MAX_PATH];
-	// strcpy(tmp_template, "/tmp/minarch-XXXXXX");
-	
-	mkdir("/tmp/nextarch",0777);
-	char tmp_dirname[255];
-	snprintf(tmp_dirname, sizeof(tmp_dirname), "%s/%s", "/tmp/nextarch",core.tag);
-	mkdir(tmp_dirname,0777);
-
-	int i, len;
-	int fd;
-	struct zip_file *zf;
-	struct zip_stat sb;
-	long long sum;
-	for (i = 0; i < zip_get_num_entries(za, 0); i++) {
-		if (zip_stat_index(za, i, 0, &sb) == 0) {
-			len = strlen(sb.name);
-			//LOG_info("Name: [%s], ", sb.name);
-			//LOG_info("Size: [%llu], ", sb.size);
-			//LOG_info("mtime: [%u]\n", (unsigned int)sb.mtime);
-			if (sb.name[len - 1] == '/') {
-				sprintf(game.tmp_path, "%s/%s", tmp_dirname, basename((char*)sb.name));
-			} else {
-				int found = 0;
-				char extension[8];
-				for (int e=0; extensions[e]; e++) {
-					sprintf(extension, ".%s", extensions[e]);
-					if (suffixMatch(extension, sb.name)) {
-						found = 1;
-						break;
-					}
-				}
-				if (!found) continue;
-
-				sprintf(game.tmp_path, "%s/%s", tmp_dirname, basename((char*)sb.name));
-				
-				// Check if file already exists and has the correct size
-				struct stat st;
-				if (stat(game.tmp_path, &st) == 0 && st.st_size == sb.size) {
-					// File already exists with correct size, skip extraction
-					LOG_info("File already exists with correct size, skipping extraction: %s\n", game.tmp_path);
-					return 1;
-				}
-				
-				zf = zip_fopen_index(za, i, 0);
-				if (!zf) {
-					LOG_error( "zip_fopen_index failed\n");
-					return 0;
-				}
-
-				// Try to create file exclusively first to avoid race condition
-				fd = open(game.tmp_path, O_RDWR | O_CREAT | O_EXCL, 0644);
-				if (fd < 0) {
-					if (errno == EEXIST) {
-						// File was created by another process, verify it's complete
-						zip_fclose(zf);
-						if (stat(game.tmp_path, &st) == 0 && st.st_size == sb.size) {
-							LOG_info("File was created by another process, using it: %s\n", game.tmp_path);
-							return 1;
-						}
-						// File exists but wrong size, try to truncate and rewrite
-						fd = open(game.tmp_path, O_RDWR | O_TRUNC, 0644);
-						if (fd < 0) {
-							LOG_error("open failed after EEXIST: %s\n", strerror(errno));
-							return 0;
-						}
-						zf = zip_fopen_index(za, i, 0);
-						if (!zf) {
-							LOG_error("zip_fopen_index failed on retry\n");
-							close(fd);
-							return 0;
-						}
-					} else {
-						LOG_error("open failed: %s\n", strerror(errno));
-						zip_fclose(zf);
-						return 0;
-					}
-				}
-				//LOG_info("Writing: %s\n", game.tmp_path);
-
-				sum = 0;
-				while (sum != sb.size) {
-					len = zip_fread(zf, buf, 100);
-					if (len < 0) {
-						LOG_error( "zip_fread failed\n");
-						close(fd);
-						zip_fclose(zf);
-						return 0;
-					}
-					write(fd, buf, len);
-					sum += len;
-				}
-				close(fd);
-				zip_fclose(zf);
-				return 1;
-			}
-		}
-	}
-	
-	if (zip_close(za) == -1) {
-		LOG_error("can't close zip archive `%s'\n", game.path);
-		return 0;
-	}
-
-	return 0;
-}
-
-///////////////////////////////
 
 
 char* onoff_labels[] = {
