+++ workspace/all/minarch/minarch_game.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void Game_open(char* path);
+void Game_close(void);
+void Game_changeDisc(char* path);
+int extract_zip(char** extensions);
